// types/bindings/Progress.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { QueryStatus } from "./QueryStatus";

export type Progress = { "status": "none" } | { "status": "intro", participant_uuid: string, } | { "status": "ongoing", participant_uuid: string, participant_name: string, view: Array<number>, query_status: QueryStatus, } | { "status": "finished", participant_uuid: string, participant_name: string, printable: boolean, };